JENKINS-30801 add unit test, update javadoc, remove unused JS
Remove 'retryBuild' JS - this is no longer used by any clients, it appears
to have been removed in a refactor in JENKINS-22210.

retryBuild functionality would retrigger the build but without any of the
upstream cause maintained - it will only work for starting a new pipeline
so nothing is really being retried.